+++ packages/moleculer-db-adapter-mongo/test/unit/index.spec.js
@@ -9,9 +9,10 @@ const mongodb = require("mongodb");
 const MongoClient = mongodb.MongoClient;
 
 function protectReject(err) {
-	if (err && err.stack)
-		console.error(err.stack);
-
+	if (err && err.stack) {
+		console.error(err);
+		console.error(err.stack);	
+	}
 	expect(err).toBe(true);
 }
