+++ packages/moleculer-db-adapter-mongoose/test/unit/index.spec.js
@@ -5,7 +5,10 @@ const MongooseStoreAdapter = require("../../src");
 const mongoose = require("mongoose");
 
 function protectReject(err) {
-	console.error(err.stack);
+	if (err && err.stack) {
+		console.error(err);
+		console.error(err.stack);	
+	}
 	expect(err).toBe(true);
 }
