+++ packages/moleculer-db/examples/simple/index.js
@@ -63,6 +63,7 @@ function start() {
 // --- TEST CASES ---
 
 let id;
+let date = new Date();
 
 // Count of posts
 checker.add("COUNT", () => broker.call("posts.count"), res => {
@@ -70,7 +71,6 @@ checker.add("COUNT", () => broker.call("posts.count"), res => {
 	return res == 0;
 });
 
-let date = new Date();
 // Create new Posts
 checker.add("--- CREATE ---", () => broker.call("posts.create", { title: "Hello", content: "Post content", votes: 2, createdAt: date, status: true }), doc => {
 	id = doc._id;
